refector: change main function and remove hooks
