+++ src/renderer/src/utils/index.test.ts
@@ -1,6 +1,6 @@
 import Big from 'big.js'
 
-import { useCalculator } from '../hooks'
+import { handleInput } from '.'
 import { pipe } from '@mobily/ts-belt'
 
 describe('[Calulator]', () => {
@@ -8,27 +8,21 @@ describe('[Calulator]', () => {
     const A = 1234
     const B = 2345
 
-    const calulator = useCalculator()
-
-    expect(pipe('=', calulator(`${A}+${B}`))).toBe(Big(A).plus(Big(B)).toString())
+    expect(pipe('=', handleInput(`${A}+${B}`))).toBe(Big(A).plus(Big(B)).toString())
   })
 
   test('Testing Oprator Divide', async () => {
     const A = 1111
     const B = 4444
 
-    const calulator = useCalculator()
-
-    expect(pipe('=', calulator(`${B}/${A}`))).toBe(Big(B).div(Big(A)).toString())
+    expect(pipe('=', handleInput(`${B}/${A}`))).toBe(Big(B).div(Big(A)).toString())
   })
 
   test('Testing Oprator Times', async () => {
     const A = 1234
     const B = 2345
 
-    const calulator = useCalculator()
-
-    expect(pipe('=', calulator(`${B}*${A}`))).toBe(Big(B).times(Big(A)).toString())
+    expect(pipe('=', handleInput(`${B}*${A}`))).toBe(Big(B).times(Big(A)).toString())
   })
 
   test('Testing Oprator Times and Divide', async () => {
@@ -36,9 +30,7 @@ describe('[Calulator]', () => {
     const B = 2222
     const C = 1234
 
-    const calulator = useCalculator()
-
-    expect(pipe('=', calulator(`${B}/${A}*${C}`))).toBe(Big(B).div(Big(A)).times(C).toString())
+    expect(pipe('=', handleInput(`${B}/${A}*${C}`))).toBe(Big(B).div(Big(A)).times(C).toString())
   })
 
   test('Testing Oprator Times and Divide and Plus(decimal)', async () => {
@@ -47,8 +39,7 @@ describe('[Calulator]', () => {
     const C = 1234
     const D = 2333.12
 
-    const calulator = useCalculator()
-    expect(pipe('=', calulator(`${B}/${A}*${C}+${D}`))).toBe(
+    expect(pipe('=', handleInput(`${B}/${A}*${C}+${D}`))).toBe(
       Big(B).div(Big(A)).times(C).plus(D).toString()
     )
   })
